handle setting hour for cvops

+++ weather-update/src/main/java/com/marlan/weatherupdate/service/missioneditor/values/Time.java
@@ -4,11 +4,11 @@ import lombok.Data;
 
 @Data
 public class Time {
-    private final int hour;
+    private final float hour;
     private final int day;
     private final int month;
 
-    public Time(int hour, int day, int month) {
+    public Time(float hour, int day, int month) {
         this.hour = hour;
         this.day = day;
         this.month = month;
